Fixed int vs unsigned int error
git-svn-id: https://vtr-verilog-to-routing.googlecode.com/svn/trunk@3249 8e3573b8-cf2c-4f14-ef6d-137439e28b8b

+++ vpr/SRC/base/check_netlist.c
@@ -73,7 +73,7 @@ void check_netlist() {
 	free_hash_table(net_hash_table);
 
 	/* Check that each block makes sense. */
-	for (i = 0; i < num_blocks; i++) {
+	for (i = 0; i < (unsigned int)num_blocks; i++) {
 		num_conn = get_num_conn(i);
 		error += check_clb_conn(i, num_conn);
 		error += check_clb_internal_nets(i);
